+++ test/e2e/framework/framework.go
@@ -162,17 +162,7 @@ func New(kubeConfigPath string) (*Framework, CleanUpFunc, error) {
 		kubeConfigPath:            kubeConfigPath,
 		SchedulingClient:          schedulingClient,
 		OpenShiftMonitoringClient: osmclient,
-		ManifestsFactory: manifests.NewFactory(
-			namespaceName,
-			userWorkloadNamespaceName,
-			nil,
-			nil,
-			nil,
-			manifests.NewAssets("../../assets"),
-			&manifests.APIServerConfig{},
-			&configv1.Console{},
-			nil,
-		),
+		ManifestsFactory:          manifests.NewFactory(namespaceName, userWorkloadNamespaceName, nil, nil, nil, manifests.NewAssets("../../assets"), &manifests.APIServerConfig{}, &configv1.Console{}),
 	}
 
 	cleanUp, err := f.setup()